### 2022-01-21 #4090 @whedon
[PRE REVIEW]: yadg: yet another datagram
**Submitting author:** @PeterKraus (<a href="http://orcid.org/0000-0002-4359-5003">Peter Kraus</a>)
**Repository:** <a href="https://github.com/dgbowl/yadg" target ="_blank">https://github.com/dgbowl/yadg</a>
**Version:** v4.0.0
