### 2022-01-22 #4090 @PeterKraus
Hi @kyleniemeyer, no worries, I understand the delays. I had a look at the reviewer list, and here are some suggestions, as requested starting at the bottom of the list:

- synthaze (general python tools)
- 1mikegrn (material science & chemistry & cheminformatics)
- varlottaang (data science & data ethics)
- bobturneruk (data engineering & good software engineering practices)
- alberto-battistel (impedance spectroscopy, electrochemistry, data analysis)
- lscalfi (electrochemistry, physical chemistry)
- mdmurbach (electrochemistry)

Cheers!
