### 2022-02-12 #4090 @PeterKraus
Hi @gallesiorob, pleasure to have you as an editor. I've listed a few names of folks who seem to have the right interest & coding preferences from the lower parts of the "signed-up reviewers" sheet [here](https://github.com/openjournals/joss-reviews/issues/4090#issuecomment-1019274321) - not sure if you want me to actually tag them or what's the process?
