+++ tests/test_electrochem.py
@@ -443,8 +443,6 @@ def test_compare_raw_values_eis_traces(input, refpath, datadir):
     with open(refpath, "r") as infile:
         ref = json.load(infile)["steps"][0]["data"][0]["raw"]["traces"]
     ret = datagram_from_input(input, "electrochem", datadir)
-    with open(r"C:\Users\krpe\yadg\tests\geis_data.json", "w") as of:
-        json.dump(ret, of)
     ret = ret["steps"][0]["data"][0]["raw"]["traces"]
     for n, trace in ret.items():
         for ax in trace:
