### 2022-02-14 aa708a1
Split PEIS/GEIS traces from cycles into timesteps. (#18)
* Fixed traces in electrochemistry files. Tests pass.

* Bump to version number.

* Removed save from test.
