### 2022-02-14 54bde8e
Fix impedance accuracy (#19)
* fix accuracy.

* Fixed tests.

* Fixed docs.

Co-authored-by: Peter Kraus <peter.kraus@geeksonbikes.net>
