+++ src/yadg/parsers/electrochem/__init__.py
@@ -3,7 +3,9 @@
 
 List of supported file formats:
 
- - EC-Lab raw data binary file and parameter settings :mod:`yadg.parsers.electrochem.eclabmpr`
- - EC-Lab human-readable text export of data :mod:`yadg.parsers.electrochem.eclabmpt`
+ - EC-Lab raw data binary file and parameter settings 
+   :mod:`yadg.parsers.electrochem.eclabmpr`
+ - EC-Lab human-readable text export of data 
+   :mod:`yadg.parsers.electrochem.eclabmpt`
 """
 from .main import process, version
