### 2022-02-15 #4166 @whedon
[REVIEW]: yadg: yet another datagram
**Submitting author:** @PeterKraus (<a href="http://orcid.org/0000-0002-4359-5003">Peter Kraus</a>)
**Repository:** https://github.com/dgbowl/yadg
**Branch with paper.md** (empty if default branch):
**Version:** v4.0.0

@aozorahime, @pythonpanda2 & @1mikegrn, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
