### 2022-02-22 #4166 @1mikegrn
Overall I feel like this is a very well put together package - most of my critiques are minor things like unused imports and whatnot. I don't have any major concerns for which I would expect any sort of revision. 👍🏻 I also noticed that there's some CI/CD built in from gitlab, I don't use gitlab so I didn't verify any of that, but I was able to easily install the package and run the test suite in a generic docker container, so I'd call it close enough 🤷🏻‍♂️.
