### 2022-02-22 #4166 @PeterKraus
> Overall I feel like this is a very well put together package - most of my critiques are minor things like unused imports and whatnot. I don't have any major concerns for which I would expect any sort of revision. 👍🏻 I also noticed that there's some CI/CD built in from gitlab, I don't use gitlab so I didn't verify any of that, but I was able to easily install the package and run the test suite in a generic docker container, so I'd call it close enough 🤷🏻‍♂️.

We've moved the package from our internal gitlab server back to github fairly recently. The gitlab repo is now archived and only the main repo on here is used. I forgot to remove the gitlab CI files after the migration. The github actions now replicate the CI functionality (with testing of every PR on Windows and Linux), including a docs deployment after every commit to master. If you have any suggestions how to improve this, I'm all ears!
