### 2022-02-22 89ad0ab
Added [testing] to setup.py.
