### 2022-02-22 d82ab9c
Let's verbose print what's installed.
