+++ .github/workflows/pull-request-commit.yml
@@ -56,7 +56,7 @@ jobs:
       - uses: actions/download-artifact@v2
         with:
           name: dist-${{ matrix.os }}-${{ matrix.pyver }}
-      - run: pip install yadg*whl
+      - run: pip install --pre --find-links=. yadg
         shell: bash
       - uses: ./.github/workflows/docs-job
       - uses: actions/upload-artifact@v2
