+++ tests/test_encoding.py
@@ -1,10 +1,5 @@
 import pytest
-from tests.utils import (
-    datagram_from_input,
-    standard_datagram_test,
-    pars_datagram_test,
-    datadir,
-)
+from tests.utils import datagram_from_input, standard_datagram_test
 
 
 @pytest.mark.parametrize(
