### 2022-02-22 4fc5040
fix all flake8 in tests except E501 and W503.
