+++ tests/test_meascsv.py
@@ -1,4 +1,3 @@
-import pytest
 from tests.utils import (
     datagram_from_input,
     standard_datagram_test,
