+++ tests/test_dummy.py
@@ -1,7 +1,6 @@
 import pytest
 import os
 import json
-import pytest
 import yadg.core
 
 from tests.schemas import dummy_1, dummy_2, dummy_3, dummy_4, dummy_5
