+++ src/yadg/parsers/qftrace/main.py
@@ -1,13 +1,7 @@
-import os
 import logging
 from scipy.signal import find_peaks
 import numpy as np
 
-import uncertainties as uc
-from uncertainties import unumpy as unp
-
-
-import yadg.dgutils
 from yadg.parsers.qftrace import fit, prune, labviewcsv
 
 version = "4.0.0"
