+++ src/yadg/parsers/chromtrace/__init__.py
@@ -5,9 +5,9 @@ List of supported file formats:
 
  - EZ-Chrom ASCII export (dat.asc) :mod:`yadg.parsers.chromtrace.ezchromasc`
  - Agilent Chemstation Chromtab (csv) :mod:`yadg.parsers.chromtrace.agilentcsv`
- - Agilent OpenLab binary signal (ch) :mod:`yadg.parsers.chromtrace.agilentch` 
- - Agilent OpenLab data archive (dx) :mod:`yadg.parsers.chromtrace.agilentdx` 
- - Fusion JSON format (json) :mod:`yadg.parsers.chromtrace.fusionjson` 
+ - Agilent OpenLab binary signal (ch) :mod:`yadg.parsers.chromtrace.agilentch`
+ - Agilent OpenLab data archive (dx) :mod:`yadg.parsers.chromtrace.agilentdx`
+ - Fusion JSON format (json) :mod:`yadg.parsers.chromtrace.fusionjson`
  - Fusion zip archive (zip) :mod:`yadg.parsers.chromtrace.fusionjson`
 """
 from .main import process, version
