### 2022-02-23 0976112
Add exclude, parse more formats in drycal, handle NaN in electrochem.
