+++ src/yadg/core/spec_schema.py
@@ -169,6 +169,7 @@ schema_step["any"]["parameters"]["any"].update(
 )
 schema_step["any"]["parameters"]["any"]["tracetype"]["one"].append("ezchrom.asc")
 schema_step["any"]["parameters"]["any"]["tracetype"]["one"].append("fusion.json")
+schema_step["any"]["parameters"]["any"]["tracetype"]["one"].append("fusion.zip")
 schema_step["any"]["parameters"]["any"]["tracetype"]["one"].append("agilent.ch")
 schema_step["any"]["parameters"]["any"]["tracetype"]["one"].append("agilent.dx")
 schema_step["any"]["parameters"]["any"]["tracetype"]["one"].append("agilent.csv")
@@ -201,6 +202,9 @@ schema_step["any"]["parameters"]["any"].update(
     }
 )
 schema_step["any"]["parameters"]["any"]["filetype"]["one"].append("drycal")
+schema_step["any"]["parameters"]["any"]["filetype"]["one"].append("drycal.csv")
+schema_step["any"]["parameters"]["any"]["filetype"]["one"].append("drycal.rtf")
+schema_step["any"]["parameters"]["any"]["filetype"]["one"].append("drycal.txt")
 
 schema = {
     "type": dict,
