+++ tests/schemas.py
@@ -165,3 +165,18 @@ gctrace_chromtab = {
         },
     },
 }
+
+exclude_1 = {
+    "metadata": {"provenance": "manual", "schema_version": "4.1.0"},
+    "steps": [
+        {
+            "parser": "dummy", 
+            "import": {
+                "folders": ["."],
+                "suffix": "json",
+                "contains": "schema",
+                "exclude": "3.1.0"
+            }
+        }
+    ],
+}
