+++ src/yadg/parsers/chromtrace/fusionzip.py
@@ -57,8 +57,7 @@ def process(fn: str, encoding: str, timezone: str) -> tuple[list, dict]:
         zf.extractall(tempdir)
         chroms = []
         meta = {}
-        common = None
-        for ffn in os.listdir(tempdir):
+        for ffn in sorted(os.listdir(tempdir)):
             ffn = os.path.join(tempdir, ffn)
             if ffn.endswith("fusion-data"):
                 _chrom, _meta = processjson(ffn, encoding, timezone)
