### 2022-02-23 77a8aee
Restructured derived data in chromtrace.
