### 2022-02-23 40e6ce5
Minor changes and patches (#24)
* Add exclude, parse more formats in drycal, handle NaN in electrochem.

* added fusion.zip

* exclude test

* exclude docs

* sort tmpdir.

* Calibration & test update.

* Restructured derived data in chromtrace.

* Fixed docs

* Accidentaly checked in autogenerated files.

* forgot this one.

* Added version and fixed tests

Co-authored-by: Peter Kraus <peter.kraus@geeksonbikes.net>
