### 2022-02-25 9f7833a
Fix tests and allow -p.
