### 2022-02-25 aee03fe
Fix preset outfile (#28)
* Fix tests and allow -p.

* Nudge version.

* Forgot debug print.

* One more debug print and remove flaky test.
