+++ .github/workflows/docs-job/action.yml
@@ -3,8 +3,6 @@ description: "Build the documentation."
 runs:
   using: "composite"
   steps:
-    - run: pip install sphinx sphinx-rtd-theme sphinx-autodoc-typehints
-      shell: bash
     - run: sphinx-apidoc -o docs/source src/yadg -T -M
       shell: bash
     - run: sphinx-build -W -b html docs/source public
