### 2022-02-25 bafd49e
JOSS: Reviewer suggestions (#23)
* Added [testing] to setup.py.

* CI needs to install [testing]

* Let's verbose print what's installed.

* Echo & try installing

* Perhaps this also works?

* The previous version installs 4.0.1 ...

* find links?

* no index

* pre for pre-release?

* E401

* Move fixtures to conftest.py, resolve F401 in tests.

* fix all flake8 in tests except E501 and W503.

* flake8 for src

* Version in specs

* versions in schema & dg.

* docs target.
