+++ src/yadg/parsers/chromtrace/ezchromasc.py
@@ -1,7 +1,7 @@
 """
 File parser for EZ-Chrom ASCII export files (dat.asc).
 
-This file format includes one timestep with multiple traces in each ASCII file. It 
+This file format includes one timestep with multiple traces in each ASCII file. It
 contains a header section, and a sequence of Y datapoints for each detector. The X
 axis is uniform between traces, and its units have to be deduced from the header.
 
