+++ src/yadg/core/process.py
@@ -17,6 +17,8 @@ from yadg.parsers import (
 import yadg.dgutils
 import yadg.core
 
+logger = logging.getLogger(__name__)
+
 
 def _infer_datagram_handler(parser: str) -> tuple[Callable, str]:
     """
@@ -129,17 +131,15 @@ def process_schema(schema: Union[list, tuple]) -> dict:
     for step in schema["steps"]:
         metadata = dict()
         timesteps = list()
-        logging.info(f'process_schema: processing step {schema["steps"].index(step)}:')
+        logger.info("processing step %d:", schema["steps"].index(step))
         handler, parserversion = _infer_datagram_handler(step["parser"])
         metadata["tag"] = step.get("tag", f"{schema['steps'].index(step):02d}")
         metadata["parser"] = {step["parser"]: {"version": parserversion}}
         todofiles = _infer_todo_files(step["import"])
         if len(todofiles) == 0:
-            logging.warning(
-                f"process_schema: No files processed by step {metadata['tag']}"
-            )
+            logger.warning("No files processed by step '%s'.", metadata["tag"])
         for tf in todofiles:
-            logging.debug(f"process_schema: processing item {tf}")
+            logger.debug("Processing item '%s'.", tf)
             ts, meta, fulldate = handler(
                 tf,
                 encoding=step["import"].get("encoding", "utf-8"),
@@ -153,11 +153,11 @@ def process_schema(schema: Union[list, tuple]) -> dict:
                 )
             assert isinstance(
                 ts, list
-            ), f"process_schema: Handler for {step['parser']} yields timesteps that are not a enclosed in a `list`."
+            ), f"Handler for {step['parser']} yields timesteps that are not a enclosed in a 'list'."
             timesteps += ts
             assert (
                 isinstance(meta, dict) or meta is None
-            ), f"process_schema: Handler for {step['parser']} yields metadata that are not a enclosed in a `dict`."
+            ), f"Handler for {step['parser']} yields metadata that are not a enclosed in a 'dict'."
             if meta is not None:
                 metadata.update(meta)
 
