+++ _sources/yadg.parsers.rst.txt
@@ -1,47 +1,0 @@
-yadg.parsers package
-====================
-
-.. automodule:: yadg.parsers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   yadg.parsers.chromtrace
-   yadg.parsers.electrochem
-   yadg.parsers.flowdata
-   yadg.parsers.masstrace
-   yadg.parsers.qftrace
-   yadg.parsers.xpstrace
-
-Submodules
-----------
-
-yadg.parsers.basiccsv module
-----------------------------
-
-.. automodule:: yadg.parsers.basiccsv
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-yadg.parsers.dummy module
--------------------------
-
-.. automodule:: yadg.parsers.dummy
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-yadg.parsers.meascsv module
----------------------------
-
-.. automodule:: yadg.parsers.meascsv
-   :members:
-   :undoc-members:
-   :show-inheritance:
