+++ _sources/version.rst.txt
@@ -1,18 +1,0 @@
-**yadg** version history
-------------------------
-
-.. include:: version.4_1.rst
-
-.. include:: version.4_0.rst
-
-.. include:: version.3_1.rst
-
-
-.. [Kraus2021x] Kraus, Wolf, Prinz, Bellini, Trunschke, Schlögl. 
-   *Towards automation of operando experiments: A case study in contactless 
-   conductivity measurements*, **2021**, 
-   DOI: https://doi.org/10.33774/chemrxiv-2021-mh17g
-
-.. [Kajfez1994] Kajfez.
-   *Linear fractional curve fitting for measurement of high Q factors*, **1994**,
-   DOI: https://doi.org/10.1109/22.299749

[…]
