### 2022-02-25 a7fc70e
Docs to public/master (#30)
* Push docs to public/master not public

* dynamic version
