+++ .github/workflows/docs-job/action.yml
@@ -5,5 +5,5 @@ runs:
   steps:
     - run: sphinx-apidoc -o docs/source src/yadg -T -M
       shell: bash
-    - run: sphinx-build -W -b html docs/source public
+    - run: sphinx-build -W -b html docs/source public/master
       shell: bash
