+++ versions/4.0.0/_sources/parsers.electrochem.rst.txt
@@ -1,159 +1,0 @@
-``electrochem``: Electrochemistry data parser
-=============================================
-The ``electrochem`` parser handles the reading and processing of files from
-BioLogic's EC-Lab, namely the binary ``.mpr`` files and the ``.mpt``
-ASCII data files.
-
-.. note::
-
-    This interface is not yet final and will change with version 5.0.0
-
-``.mpt`` files are made up of a header portion (with the technique
-parameter sequences and an optional loops section) and a tab-separated
-data table.
-
-``.mpr`` files are structured in a set of "modules", one concerning
-settings, one for actual data, one for logs, and an optional loops
-module. The parameter sequences can be found in the settings module.
-
-The basic function of the parser is to:
-
-#. Read in the technique data and create timesteps.
-#. Collect ``metadata`` such as the measurement settings and the loops
-   contained in a given file.
-#. Collect ``common`` data describing the technique parameter sequences.
-
-These are the implemented techniques for which the technique parameter
-sequences can be parsed:
-
-+------+-------------------------------------------------+
-| CA   | Chronoamperometry / Chronocoulometry            |
-+------+-------------------------------------------------+
-| CP   | Chronopotentiometry                             |
-+------+-------------------------------------------------+
-| CV   | Cyclic Voltammetry                              |
-+------+-------------------------------------------------+
-| GCPL | Galvanostatic Cycling with Potential Limitation |
-+------+-------------------------------------------------+
-| GEIS | Galvano Electrochemical Impedance Spectroscopy  |
-+------+-------------------------------------------------+
-| LOOP | Loop                                            |
-+------+-------------------------------------------------+
-| LSV  | Linear Sweep Voltammetry                        |
-+------+-------------------------------------------------+
-| MB   | Modulo Bat                                      |
-+------+-------------------------------------------------+
-| OCV  | Open Circuit Voltage                            |
-+------+-------------------------------------------------+
-| PEIS | Potentio Electrochemical Impedance Spectroscopy |
-+------+-------------------------------------------------+
-| WAIT | Wait                                            |
-+------+-------------------------------------------------+
-| ZIR  | IR compensation (PEIS)                          |
-+------+-------------------------------------------------+
-
-Usage
------
-The use of ``electrochem`` can be specified using the ``"parser"`` keyword in
-the `schema`. Further information can be specified in the
-``"parameters"`` :class:`(dict)`:
-
-- ``"filetype"`` :class:`(str)`: The file type of the raw data file.
-  See :ref:`here<parsers_electrochem_formats>` for details.
-
-.. _parsers_electrochem_provides:
-
-Provides
---------
-The ``electrochem`` parser loads all the columns present in an ``.mpt`` or
-``.mpr`` file.
-
-.. note::
-
-    ``.mpt`` files can contain more data than the corresponding binary
-    ``.mpr`` file.
-
-Most techniques write data that can be understood as a series of
-timesteps. Each timestep provided by the parser has the following
-format:
-
-.. code-block:: yaml
-
-    - fn   !!str
-    - uts  !!float
-    - raw:
-        "{{ col1 }}":  !!int
-        "{{ col2 }}":
-          {n: !!float, s: !!float, u: !!str}
-
-For impedance spectroscopy techniques (PEIS, GEIS), the data is made up
-of spectroscopy traces. The data is thus split into traces by the column
-``"cycle number"`` and cast into a single timestep. Each trace now
-corresponds to a spectroscopy scan, indexed by ``"cycle number"`` / 
[…]
