+++ versions/4.0.1/_sources/yadg.parsers.electrochem.rst.txt
@@ -1,42 +1,0 @@
-yadg.parsers.electrochem package
-================================
-
-.. automodule:: yadg.parsers.electrochem
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-yadg.parsers.electrochem.eclabmpr module
-----------------------------------------
-
-.. automodule:: yadg.parsers.electrochem.eclabmpr
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-yadg.parsers.electrochem.eclabmpt module
-----------------------------------------
-
-.. automodule:: yadg.parsers.electrochem.eclabmpt
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-yadg.parsers.electrochem.eclabtechniques module
------------------------------------------------
-
-.. automodule:: yadg.parsers.electrochem.eclabtechniques
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-yadg.parsers.electrochem.main module
-------------------------------------
-
-.. automodule:: yadg.parsers.electrochem.main
-   :members:
-   :undoc-members:
-   :show-inheritance:
