+++ 4.0.0/_sources/yadg.parsers.flowdata.rst.txt
@@ -1,0 +1,26 @@
+yadg.parsers.flowdata package
+=============================
+
+.. automodule:: yadg.parsers.flowdata
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+yadg.parsers.flowdata.drycal module
+-----------------------------------
+
+.. automodule:: yadg.parsers.flowdata.drycal
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+yadg.parsers.flowdata.main module
+---------------------------------
+
+.. automodule:: yadg.parsers.flowdata.main
+   :members:
+   :undoc-members:
+   :show-inheritance:
