+++ 4.0.1/_sources/objects.rst.txt
@@ -1,0 +1,12 @@
+
+**yadg** objects
+----------------
+
+Several objects are defined within **yadg**. Their respective documentation is
+available in at the following links:
+
+.. toctree::
+   :maxdepth: 1
+   
+   object.schema
+   object.datagram
