+++ 4.0.0/_sources/yadg.core.rst.txt
@@ -1,0 +1,42 @@
+yadg.core package
+=================
+
+.. automodule:: yadg.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+yadg.core.process module
+------------------------
+
+.. automodule:: yadg.core.process
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+yadg.core.spec\_datagram module
+-------------------------------
+
+.. automodule:: yadg.core.spec_datagram
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+yadg.core.spec\_schema module
+-----------------------------
+
+.. automodule:: yadg.core.spec_schema
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+yadg.core.validators module
+---------------------------
+
+.. automodule:: yadg.core.validators
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
