+++ .github/workflows/push-master.yml
@@ -49,7 +49,7 @@ jobs:
       - uses: actions/upload-artifact@v2
         with: 
           name: public-master
-          path: public
+          path: public/master
   deploy:
     needs: [pages]
     runs-on: ubuntu-latest
@@ -61,8 +61,8 @@ jobs:
       - uses: actions/download-artifact@v2
         with:
           name: public-master
-          path: public
+          path: public/master
       - uses: peaceiris/actions-gh-pages@v3
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: ./public
+          publish_dir: ./public/master
