+++ 4.0.1/_sources/modules.rst.txt
@@ -1,7 +1,0 @@
-**yadg** autodocs
------------------
-
-.. toctree::
-   :maxdepth: 4
-
-   yadg
