+++ 4.0.1/_sources/parsers.rst.txt
@@ -1,19 +1,0 @@
-**yadg** parsers
-----------------------
-
-**yadg** offloads the responsibility of processing the raw data files to its 
-modules, called parsers. The currently implemented parsers are documented in the 
-following links:
-
-.. toctree::
-   :maxdepth: 1
-
-   parsers.dummy
-   parsers.basiccsv
-   parsers.chromtrace
-   parsers.qftrace
-   parsers.meascsv
-   parsers.flowdata
-   parsers.masstrace
-   parsers.xpstrace
-   parsers.electrochem
