+++ 4.0.1/_sources/yadg.rst.txt
@@ -1,0 +1,36 @@
+yadg package
+============
+
+.. automodule:: yadg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   yadg.core
+   yadg.dgutils
+   yadg.parsers
+
+Submodules
+----------
+
+yadg.main module
+----------------
+
+.. automodule:: yadg.main
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+yadg.subcommands module
+-----------------------
+
+.. automodule:: yadg.subcommands
+   :members:
+   :undoc-members:
+   :show-inheritance:
