+++ master/_sources/parsers.dummy.rst.txt
@@ -1,0 +1,4 @@
+``dummy``: A dummy parser
+`````````````````````````
+This is a dummy parser, used mainly for testing. It provides no real functionality.
+See also :mod:`yadg.parsers.dummy`.
