+++ .github/workflows/push-master.yml
@@ -65,4 +65,5 @@ jobs:
       - uses: peaceiris/actions-gh-pages@v3
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: ./public/master
+          publish_dir: ./public
+          keep_files: true
