+++ docs/source/conf.py
@@ -41,7 +41,7 @@ extensions = [
 ]
 
 # Add any paths that contain templates here, relative to this directory.
-# templates_path = ['_templates']
+templates_path = ['_templates']
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -72,3 +72,11 @@ html_static_path = ["images"]
 
 autosummary_generate = True
 autodoc_default_flags = ["members", "undoc-members", "show-inheritance"]
+
+html_context = {
+    'versions': [
+        "master",
+        "4.0.1", 
+        "4.0.0"
+    ]
+}
