+++ 4.0.0/_sources/parsers.dummy.rst.txt
@@ -1,4 +1,0 @@
-``dummy``: A dummy parser
-`````````````````````````
-This is a dummy parser, used mainly for testing. It provides no real functionality.
-See also :mod:`yadg.parsers.dummy`.
