+++ 4.0.1/_sources/object.datagram.rst.txt
@@ -1,147 +1,0 @@
-.. _object_datagram:
-
-What is a `datagram`
-````````````````````
-The `datagram` is a structured and annotated representation of both raw and 
-processed data. Here, "raw data" corresponds to data present in the output files
-as they come out of an instrument directly, while "processed data" corresponds 
-to data after any processing -- whether the processing is applying a calibration 
-curve, or a more involved transformation, such as deriving :math:`Q_0` and 
-:math:`f_0` from :math:`\Gamma(f)` in the :mod:`yadg.parsers.qftrace` module. 
-
-The `datagram` is designed to be a mirror of the raw data files, with:
-
-    - uncertainties of measured datapoints
-    - units associated with data
-    - a consistent data structure for timestamped traces
-    - a consistent variable mapping between different filetypes
-
-Additionally, the `datagram` is  annotated by relevant metadata, including:
-
-    - version information
-    - clear provenance of raw data 
-    - documentation of any post-processing
-    - uniform data timestamping between all `datagrams`
-
-.. note::
-    The datagram does not guarantee that timesteps are normalized. This means entries 
-    may or may not be present in all timesteps within a step. An example of this would
-    be if an analyte appears in chromatographic traces after the first timesteps - the 
-    entry corresponding to the concentration of that analyte is **not** back-filled by
-    **yadg**.
-
-The basic structure of a `datagram` is as follows:
-
-.. code-block:: json
-
-    {
-        "metadata": {
-            "yadg": {"version": "4.0.0", "command": "pytest"},
-            "date": "2021-09-29 09:20:00",
-            "input_schema": {"..."},
-            "datagram_version": "4.0.0"
-        },
-        "steps": [
-            {
-                "metadata": {
-                    "tag": "flow",
-                    "parser": {"meascsv": {"version": "4.0.0"}}
-                },
-                "data": [
-                    {
-                        "uts": 1632900000.0,
-                        "fn": "foo.csv", 
-                        "raw": {
-                            "flow": {"n": 15.0, "s": 0.1, "u": "ml/min"},                    
-                            "C3H8": {"n": 0.0305, "s": 0.0010, "u": " "}, 
-                            "O2": {"n": 0.0895, "s": 0.0010, "u": " "}, 
-                            "N2": {"n": 0.8800, "s": 0.0100, "u": " "}
-                        }
-                    },{
-                        "uts": 1632900060.0,
-                        "fn": "foo.csv", 
-                        "raw": {
-                            "flow": {"n": 14.9, "s": 0.1, "u": "ml/min"},                    
-                            "C3H8": {"n": 0.0304, "s": 0.0010, "u": " "}, 
-                            "O2": {"n": 0.0896, "s": 0.0010, "u": " "}, 
-                            "N2": {"n": 0.8800, "s": 0.0100, "u": " "}
-                        }
-                    },{
-                        "uts": 1632900120.0,
-                        "fn": "foo.csv",
-                        "raw": {
-                            "flow": {"n": 15.0, "s": 0.1, "u": "ml/min"},                    
-                            "C3H8": {"n": 0.0305, "s": 0.0010, "u": " "}, 
-                            "O2": {"n": 0.0900, "s": 0.0010, "u": " "}, 
-                            "N2": {"n": 0.8795, "s": 0.0100, "u": " "}
-                    },{
-                        "uts": 1632900180.0,
-                        "fn": "foo.csv",
-                        "raw": {
-                        "flow": {"n": 15.0, "s": 0.1, "u": "ml/min"},                    
[…]
