+++ 4.0.0/_sources/modules.rst.txt
@@ -1,0 +1,7 @@
+**yadg** autodocs
+-----------------
+
+.. toctree::
+   :maxdepth: 4
+
+   yadg
