### 2022-03-02 4d8c2dd
Fix typos in chromtrace logging strings
gctrace should be chromtrace, see diff.
