+++ src/yadg/parsers/electrochem/eclabtechniques.py
@@ -34,6 +34,8 @@ import bisect
 import math
 import logging
 
+logger = logging.getLogger(__name__)
+
 # Short helper function for constructing params.
 def _prepend_ns(settings: list[str], params: list) -> list[str]:
     """Prepends the 'Ns' parameter to the parameters if present.
@@ -1067,7 +1069,7 @@ def get_resolution(name: str, value: float, Erange: float, Irange: float) -> flo
     elif name in ["I", "|I|"]:
         # VMP-3: 0.004% of FSR
         if Irange is None:
-            logging.warning("get_resolution: 'Irange' not specified. Using 'I'.")
+            logger.warning("'I range' not specified. Using 'I'.")
             Irange = 10 ** math.ceil(math.log10(value))
         return Irange * 0.004 / 100
     elif name in ["freq"]:
