+++ src/yadg/parsers/electrochem/eclabtechniques.py
@@ -27,7 +27,6 @@ in :func:`get_resolution`.
 
 """
 import re
-
 import numpy as np
 from typing import Union
 import bisect
