+++ src/yadg/parsers/electrochem/eclabmpt.py
@@ -127,12 +127,8 @@ column_units = {
     "R/Ohm": ("R", "Ω"),
     "Rcmp/Ohm": ("Rcmp", "Ω"),
     "Re(Y)/Ohm-1": ("Re(Y)", "S"),
-    "Re(Y)/Ohm-1": ("Re(Y)", "S"),
-    "Re(Z)/Ohm": ("Re(Z)", "Ω"),
     "Re(Z)/Ohm": ("Re(Z)", "Ω"),
     "Re(Zce)/Ohm": ("Re(Zce)", "Ω"),
-    "Re(Zce)/Ohm": ("Re(Zce)", "Ω"),
-    "Re(Zwe-ce)/Ohm": ("Re(Zwe-ce)", "Ω"),
     "Re(Zwe-ce)/Ohm": ("Re(Zwe-ce)", "Ω"),
     "step time/s": ("step time", "s"),
     "THD Ewe/%": ("THD Ewe", "%"),
