### 2022-03-02 b32af58
Implement module-level logger (#29)
* main

* dgutils.

* Replaced all logging with logger.

* Relative imports.

* lgtm changes.
