### 2022-03-02 243c964
Fix type hinting
