### 2022-03-03 e426a81
Fix typo in docs and update type hints (#39)
* Fix typos in chromtrace logging strings

gctrace should be chromtrace, see diff.

* Fix type hinting

* Fix incomplete merge
