+++ src/yadg/parsers/chromtrace/fusionjson.py
@@ -29,7 +29,7 @@ import numpy as np
 from ...dgutils.dateutils import str_to_uts
 
 
-def process(fn: str, encoding: str, timezone: str) -> tuple[list, dict, dict]:
+def process(fn: str, encoding: str, timezone: str) -> tuple[list, dict]:
     """
     Fusion json format.
 
