### 2022-03-03 #4166 @pythonpanda2
@galessiorob  Not sure if this is an issue due to Safari. I still cannot select the tick option in check list. I am going to try again in Firefox.
