### 2022-03-04 7b3724e
Fix for external devices in mpr files. (#40)
