### 2022-03-04 44e4bcb
Let's start with no smoothing case.
