+++ tests/test_gctrace.py
@@ -1,5 +1,7 @@
 import pytest
 import os
+import pickle
+import numpy as np
 from tests.schemas import gctrace_chromtab
 from tests.utils import (
     datagram_from_input,
@@ -218,78 +220,86 @@ def test_datagram_from_gctrace(input, ts, datadir):
                 },
             },
         ),
-        (
-            {  # ts1 - fusion zip file, minimal smoothing
-                "folders": ["."],
-                "prefix": "",
-                "suffix": "zip",
-                "parameters": {
-                    "tracetype": "fusion.zip",
-                    "calfile": "calibrations/2022-02-04-GCcal.json",
-                    "detectors": {
-                        "TCD1": {
-                            "id": 0,
-                            "peakdetect": {
-                                "window": 3,
-                                "polyorder": 2,
-                                "prominence": 10.0,
-                                "threshold": 10.0
-                            }
-                        }
-                    }
-                },
-            },
-            {
-                "nsteps": 1,
-                "step": 0,
-                "nrows": 12,
-                "method": "AS_Cal_20220204",
-                "point": 4,
-                "xout": {
-                    "H2":  {"n": 0.6220791, "s": 0.0131074, "u": " "},
-                    "CH4": {"n": 0.0578758, "s": 0.0015254, "u": " "},
-                    "CO":  {"n": 0.0136661, "s": 0.0202863, "u": " "},
-                },
-            },
-        ),
-        (
-            {  # ts2 - fusion zip file, default smoothing
-                "folders": ["."],
-                "prefix": "",
-                "suffix": "zip",
-                "parameters": {
-                    "tracetype": "fusion.zip",
-                    "calfile": "calibrations/2022-02-04-GCcal.json",
-                    "detectors": {
-                        "TCD1": {
-                            "id": 0,
-                            "peakdetect": {
-                                "window": 7,
-                                "polyorder": 3,
-                                "prominence": 10.0,
-                                "threshold": 10.0
-                            }
-                        }
-                    }
-                },
-            },
-            {
-                "nsteps": 1,
-                "step": 0,
-                "nrows": 12,
-                "method": "AS_Cal_20220204",
-                "point": 4,
-                "xout": {
-                    "H2":  {"n": 0.6224006, "s": 0.0130895, "u": " "},
-                    "CH4": {"n": 0.0581437, "s": 0.0015272, "u": " "},
-                    "CO":  {"n": 0.0136409, "s": 0.0202494, "u": " "},
-                },
-            },
-        ),
+        #(
+        #    {  # ts1 - fusion zip file, minimal smoothing
+        #        "folders": ["."],
+        #        "prefix": "",
+        #        "suffix": "zip",
+        #        "parameters": {
+        #            "tracetype": "fusion.zip",
+        #            "calfile": "calibrations/2022-02-04-GCcal.json",
+        #            "detectors": {
+        #                "TCD1": {
+        #                    "id": 0,
+        #                    "peakdetect": {
+        #                        "window": 3,
+        #                        "polyorder": 2,
+        #                        "prominence": 10.0,
+        #                        "threshold": 10.0
+        #                    }
+        #                }
+        #            }
+        #        },
+        #    },
+        #    {
+        #        "nsteps": 1,
+        #        "step": 0,
+        #        "nrows": 12,
+        #        "method": "AS_Cal_20220204",
+        #        "point": 4,
+        #        "xout": {
[…]
