+++ tests/test_gctrace.py
@@ -295,11 +295,20 @@ def test_integration(input, ts, datadir):
     ret = datagram_from_input(input, "chromtrace", datadir)
     #standard_datagram_test(ret, ts)
     pmax = ret["steps"][0]["data"][3]["derived"]["pmax"][0]
-#    with open(r"C:\Users\krpe\yadg\ref.pkl", "wb") as ouf:
-#        pickle.dump(pmax, ouf)
-    with open(r"ref.pkl", "rb") as inf:
+    pspec = ret["steps"][0]["data"][3]["derived"]["pspec"][0]
+    #with open(r"C:\Users\krpe\yadg\tests\test_gctrace\pspec.pkl", "wb") as ouf:
+    #    pickle.dump(pspec, ouf)
+    with open(r"pmax.pkl", "rb") as inf:
         rmax = pickle.load(inf)
     for k in pmax.keys():
         print(k)
         assert np.array_equal(rmax[k], pmax[k])
+    with open(r"pspec.pkl", "rb") as inf:
+        rspec = pickle.load(inf)
+    print(pspec)
+    for i in range(len(pspec)):
+        for k in {"llim", "rlim", "max"}:
+            print(i, k)
+            assert rspec[i][k] == pspec[i][k]
+    
     special_datagram_test(ret, ts)
