+++ tests/test_gctrace.py
@@ -298,8 +298,8 @@ def test_integration(input, ts, datadir):
     pspec = ret["steps"][0]["data"][3]["derived"]["pspec"][0]
     pgrad = ret["steps"][0]["data"][3]["derived"]["pgrad"][0]
     pints = ret["steps"][0]["data"][3]["derived"]["pints"][0]
-    with open(r"C:\Users\krpe\yadg\tests\test_gctrace\pints.pkl", "wb") as ouf:
-        pickle.dump(pints, ouf)
+    #with open(r"C:\Users\krpe\yadg\tests\test_gctrace\pints.pkl", "wb") as ouf:
+    #    pickle.dump(pints, ouf)
     with open(r"pmax.pkl", "rb") as inf:
         rmax = pickle.load(inf)
     for k in pmax.keys():
@@ -325,5 +325,8 @@ def test_integration(input, ts, datadir):
             print(i, k, kk, pints[k][kk], rints[k][kk])
             assert pints[k][kk].n == rints[k][kk].n
             assert pints[k][kk].s == rints[k][kk].s
+    print(ret["steps"][0]["data"][3]["derived"]["norm"])
+    assert ret["steps"][0]["data"][3]["derived"]["norm"].n == pytest.approx(0.97650072, abs=1e-6)
+    assert ret["steps"][0]["data"][3]["derived"]["norm"].s == pytest.approx(0.01768264, abs=1e-6)
     special_datagram_test(ret, ts)
     
