### 2022-03-04 cb31781
manual normalisation
