### 2022-03-04 cbdf66e
point 4 throughout. Both traces fixed.
