### 2022-03-04 5855ccf
Let's test poor smoothing.
