+++ tests/test_gctrace.py
@@ -189,8 +189,41 @@ def test_datagram_from_gctrace(input, ts, datadir):
 @pytest.mark.parametrize(
     "input, ts",
     [
+        #(
+        #    {  # ts0 - fusion zip file, no smoothing
+        #        "folders": ["."],
+        #        "prefix": "",
+        #        "suffix": "zip",
+        #        "parameters": {
+        #            "tracetype": "fusion.zip",
+        #            "calfile": "calibrations/2022-02-04-GCcal.json",
+        #            "detectors": {
+        #                "TCD1": {
+        #                    "id": 0,
+        #                    "peakdetect": {"prominence": 10.0, "threshold": 10.0}
+        #                },
+        #                "TCD2": {
+        #                    "id": 1,
+        #                    "peakdetect": {"prominence": 10.0, "threshold": 10.0}
+        #                }
+        #            }
+        #        },
+        #    },
+        #    {
+        #        "nsteps": 1,
+        #        "step": 0,
+        #        "nrows": 12,
+        #        "method": "AS_Cal_20220204",
+        #        "point": 4,
+        #        "xout": {
+        #            "H2":  {"n": 0.6333887, "s": 0.0493695, "u": " "},
+        #            "CH4": {"n": 0.0557875, "s": 0.0044479, "u": " "},
+        #            "CO":  {"n": 0.0000000, "s": 0.0373813, "u": " "},
+        #        },
+        #    },
+        #),
         (
-            {  # ts0 - fusion zip file, no smoothing
+            {  # ts1 - fusion zip file, minimal smoothing
                 "folders": ["."],
                 "prefix": "",
                 "suffix": "zip",
@@ -200,7 +233,12 @@ def test_datagram_from_gctrace(input, ts, datadir):
                     "detectors": {
                         "TCD1": {
                             "id": 0,
-                            "peakdetect": {"prominence": 10.0, "threshold": 10.0}
+                            "peakdetect": {
+                                "window": 3,
+                                "polyorder": 2,
+                                "prominence": 10.0,
+                                "threshold": 10.0
+                            }
                         },
                         "TCD2": {
                             "id": 1,
@@ -216,46 +254,12 @@ def test_datagram_from_gctrace(input, ts, datadir):
                 "method": "AS_Cal_20220204",
                 "point": 4,
                 "xout": {
-                    "H2":  {"n": 0.6333887, "s": 0.0493695, "u": " "},
-                    "CH4": {"n": 0.0557875, "s": 0.0044479, "u": " "},
-                    "CO":  {"n": 0.0000000, "s": 0.0373813, "u": " "},
+                    "H2":  {"n": 0.6230384, "s": 0.0437859, "u": " "},
+                    "CH4": {"n": 0.0579651, "s": 0.0041758, "u": " "},
+                    "CO":  {"n": 0.0136871, "s": 0.0203378, "u": " "},
                 },
             },
         ),
-        #(
-        #    {  # ts1 - fusion zip file, minimal smoothing
-        #        "folders": ["."],
-        #        "prefix": "",
-        #        "suffix": "zip",
-        #        "parameters": {
-        #            "tracetype": "fusion.zip",
-        #            "calfile": "calibrations/2022-02-04-GCcal.json",
-        #            "detectors": {
-        #                "TCD1": {
[…]
