### 2022-03-04 1b904cb
filter hessian / gradient below 1e-10
