+++ tests/test_gctrace.py
@@ -1,8 +1,5 @@
 import pytest
 import os
-import pickle
-import numpy as np
-import uncertainties as uc
 from tests.schemas import gctrace_chromtab
 from tests.utils import (
     datagram_from_input,
@@ -200,13 +197,13 @@ def test_datagram_from_gctrace(input, ts, datadir):
                     "detectors": {
                         "TCD1": {
                             "id": 0,
-                            "peakdetect": {"prominence": 10.0, "threshold": 10.0}
+                            "peakdetect": {"prominence": 10.0, "threshold": 10.0},
                         },
                         "TCD2": {
                             "id": 1,
-                            "peakdetect": {"prominence": 10.0, "threshold": 10.0}
-                        }
-                    }
+                            "peakdetect": {"prominence": 10.0, "threshold": 10.0},
+                        },
+                    },
                 },
             },
             {
@@ -216,9 +213,9 @@ def test_datagram_from_gctrace(input, ts, datadir):
                 "method": "AS_Cal_20220204",
                 "point": 4,
                 "xout": {
-                    "H2":  {"n": 0.6333887, "s": 0.0493695, "u": " "},
+                    "H2": {"n": 0.6333887, "s": 0.0493695, "u": " "},
                     "CH4": {"n": 0.0557875, "s": 0.0044479, "u": " "},
-                    "CO":  {"n": 0.0000000, "s": 0.0373813, "u": " "},
+                    "CO": {"n": 0.0000000, "s": 0.0373813, "u": " "},
                 },
             },
         ),
@@ -237,17 +234,14 @@ def test_datagram_from_gctrace(input, ts, datadir):
                                 "window": 3,
                                 "polyorder": 2,
                                 "prominence": 10.0,
-                                "threshold": 10.0
-                            }
+                                "threshold": 10.0,
+                            },
                         },
                         "TCD2": {
                             "id": 1,
-                            "peakdetect": {
-                                "prominence": 10.0, 
-                                "threshold": 10.0
-                            }
-                        }
-                    }
+                            "peakdetect": {"prominence": 10.0, "threshold": 10.0},
+                        },
+                    },
                 },
             },
             {
@@ -257,9 +251,9 @@ def test_datagram_from_gctrace(input, ts, datadir):
                 "method": "AS_Cal_20220204",
                 "point": 4,
                 "xout": {
-                    "H2":  {"n": 0.6230303, "s": 0.0437848, "u": " "},
+                    "H2": {"n": 0.6230303, "s": 0.0437848, "u": " "},
                     "CH4": {"n": 0.0579773, "s": 0.0041758, "u": " "},
-                    "CO":  {"n": 0.0136871, "s": 0.0203378, "u": " "},
+                    "CO": {"n": 0.0136871, "s": 0.0203378, "u": " "},
                 },
             },
         ),
@@ -278,17 +272,14 @@ def test_datagram_from_gctrace(input, ts, datadir):
                                 "window": 7,
                                 "polyorder": 3,
                                 "prominence": 10.0,
[…]
