+++ src/yadg/parsers/chromtrace/integration.py
@@ -243,10 +243,6 @@ def integrate_trace(traces: dict, chromspec: dict) -> tuple[dict, dict]:
 
     peaks = {}
     comp = []
-    pmax = []
-    pspec = []
-    ygrs = []
-    ints = []
     for detname, spec in chromspec.items():
         for det in traces.keys():
             if traces[det]["id"] == spec["id"]:
@@ -261,12 +257,8 @@ def integrate_trace(traces: dict, chromspec: dict) -> tuple[dict, dict]:
         pd = spec.get("peakdetect", {})
         smooth = _get_smooth_yvals(ys[0], pd)
         peakmax, ygrad, _ = _find_peak_maxima(smooth, pd)
-        pmax.append(peakmax)
-        ygrs.append(ygrad)
         peakspec = _find_peak_edges(smooth, ygrad, peakmax, pd)
-        pspec.append(peakspec)
         integrated = _integrate_peaks(xs, ys, peakspec, spec["species"])
-        ints.append(integrated)
         peaks[detname] = {}
         for k, v in integrated.items():
             peaks[detname][k] = {
@@ -314,10 +306,5 @@ def integrate_trace(traces: dict, chromspec: dict) -> tuple[dict, dict]:
         "height": height,
         "concentration": conc,
         "peaks": peaks,
-        #"pmax": pmax,
-        #"pspec": pspec,
-        #"pgrad": ygrs,
-        #"pints": ints,
-        #"norm": norm,
     }
     return derived
