### 2022-03-04 17b888a
Fix integration reliability in chromtrace (#42)
* Establish what works.

* Let's start with no smoothing case.

* Forgot about this one.

* 3rd time lucky.

* peak spec

* Double check that pspec is correct.

* grad comp.

* More tests.

* normalisation

* store fresh ufloat.

* redo naming.

* check concs.

* add ref

* manual normalisation

* point 4

* point 4 throughout. Both traces fixed.

* Let's test poor smoothing.

* filter hessian / gradient below 1e-10

* forgot one here

* Local tests pass.

* Tidy-up.
