+++ src/yadg/parsers/chromtrace/integration.py
@@ -97,11 +97,13 @@ def _find_peak_maxima(yvals: np.ndarray, pd: dict) -> dict:
     peaks["-"] = res
     # gradient: find peaks and inflection points
     grad = np.gradient(yvals)
-    res = np.nonzero(np.diff(np.sign(grad)))[0] + 1
+    gm = 1 * abs(grad) > 1e-10
+    res = np.nonzero(np.diff(np.sign(grad * gm)))[0] + 1
     peaks["gradzero"] = res
     # second derivative: find peaks
     hess = np.gradient(grad)
-    res = np.nonzero(np.diff(np.sign(hess)))[0] + 1
+    hm = 1 * abs(hess) > 1e-10
+    res = np.nonzero(np.diff(np.sign(hess * hm)))[0] + 1
     peaks["hesszero"] = res
     return peaks, grad, hess
 
@@ -275,27 +277,29 @@ def integrate_trace(traces: dict, chromspec: dict) -> tuple[dict, dict]:
                     "n": x.n,
                     "s": x.s,
                     "u": spec["species"][k]["calib"].get("unit", "%"),
-                    "uf": x,
                 }
             if k not in comp:
                 comp.append(k)
-    xout = {}
+    xtemp = {}
     area = {}
     conc = {}
     height = {}
     for s in comp:
         for d, ds in chromspec.items():
             if s in peaks[d] and "c" in peaks[d][s]:
-                v = peaks[d][s]["c"].pop("uf")
-                if ds.get("prefer", False) or s not in xout:
-                    xout[s] = v
+                if ds.get("prefer", False) or s not in xtemp:
+                    xtemp[s] = uc.ufloat(
+                        peaks[d][s]["c"]["n"],
+                        peaks[d][s]["c"]["s"],
+                    )
                     area[s] = peaks[d][s]["A"]
                     height[s] = peaks[d][s]["h"]
                     conc[s] = peaks[d][s]["c"]
-    norm = sum([xout[k] for k in xout.keys()])
-    for s in xout:
-        xnorm = xout[s] / norm
-        xout[s] = {"n": xnorm.n, "s": xnorm.s, "u": " "}
+    norm = sum([v for k, v in xtemp.items()])
+    xout = {}
+    for k, v in xtemp.items():
+        xnorm = v / norm
+        xout[k] = {"n": xnorm.n, "s": xnorm.s, "u": " "}
     derived = {
         "xout": xout,
         "area": area,
