+++ tests/test_lctrace.py
@@ -18,7 +18,7 @@ def special_datagram_test(datagram, testspec):
 
 
 refvals = {
-    "a": {"n": 4210.9762337, "s": 3.1576578, "u": " "},
+    "a": {"n": 4203.5798851, "s": 3.15755412, "u": " "},
     "b": {"n": 23349.8855948, "s": 16.2068347, "u": " "},
     "c": {"n": 71369.5831322, "s": 44.2360720, "u": " "},
     "d": {"n": 13119.4243886, "s": 7.23123702, "u": " "},
