### 2022-03-04 #4166 @kyleniemeyer
@galessiorob @pythonpanda2  I think the issue is that since we switched to the new editorialbot infrastructure, the checklists at the top no longer work (@arfon is that right?)

Instead, @pythonpanda2 can you comment `@editorialbot generate my checklist` ? This will generate one for you that you can edit.
