### 2022-03-07 dgbowl/yadg#43 @pythonpanda2
Installation
The  `pip install yadg` worked flawlessly  on my brand new apple silicon Mac book pro.

It would be nice to list the install instruction   in the GitHub  readme docs.

A side note . The project page url listed in the readme seems to be broken : https://dgbowl.github.io/master/index.html

`404 File not found `
