### 2022-03-07 #4166 @pythonpanda2
https://github.com/dgbowl/yadg/issues/43
