### 2022-03-08 dgbowl/yadg#43 @PeterKraus
Hi @pythonpanda2, thanks for the suggestions. I'm not sure where the broken link is - I checked the `README.md` and both the `setup.py` (for pypi) and `conf.py` (for docs) and they all seem to be pointing to the correct page (not 404).  Could you please point me to it?
