### 2022-03-08 dgbowl/yadg#43 @pythonpanda2
Apologies. The `pip install` did work fine. It was just the project documentation page which was not opening yesterday , but now it seems to be working. So I will close this issue and tick off the check list. Thanks.
