### 2022-03-12 dgbowl/yadg#45 @pythonpanda2
References 
Kindly fix the following references in the article

(Drawl, 2020) Draxl, C. (2020, July). FAIRmat. 2. NFDI Conference. (Incomplete)
