### 2022-03-12 dgbowl/yadg#46 @pythonpanda2
General query 
I applaud the authors for developing a software package with a focus  on the FAIR guiding principles.  Keep in mind I have no access to any of the laboratory equipment/ data generators mentioned in the article, which mean my bring me to some of the following query

1) Is the package developed with a focus on data generated from Chromatography, and Electrochemistry at EMPA.

2)  How easy would it be for an experimentalist/ alternate set up located in another facility to use yadg ?

3)  If this is indeed the case would you be able to link up some user focused examples ?

4) How does yadg handle the non standardized intermediate step (Figure 1)? Would the use of proprietary file formats pose a challenge for  FAIR ?

5) Kindly provide a brief statement in the article regarding  how this software compares to other commonly-used packages? (JOSS requirement)
