### 2022-03-13 dgbowl/yadg#46 @PeterKraus
Dear @pythonpanda2, thank you for the comments. Let me respond to each of them below:
```
1. Is the package developed with a focus on data generated from Chromatography, and Electrochemistry at EMPA.
```
So the package is currently developed by myself at Empa, in the lab of Corsin Battaglia. Significant contributions to the code have been made by our co-author, @vetschn. However, I have started working on `yadg` a few years ago at FHI.

I wouldn't say the package focuses on chromatography and electrochemistry, but those are the two areas of chemistry where the (expensive and proprietary) instruments often produce binary data files, so that's what `yadg` currently does. But, it also supports parsing network analyser traces and their fitting to obtain quality factors.

```
2. How easy would it be for an experimentalist/ alternate set up located in another facility to use yadg ?
```
I would hope very easy. It is my understanding that `yadg` is still being used to post-process data from a microwave conductivity measurement device, that is now located at a completely independent institution (FU Berlin). We are also open to contributions from others.

```
3. If this is indeed the case would you be able to link up some user focused examples ?
```
Yes, I will make some practical examples - perhaps as a binder-ready jupyter notebook. Basic usage is described [here](https://dgbowl.github.io/yadg/master/usage.html), and the documentation of each parser is [here](https://dgbowl.github.io/yadg/master/parsers.html). I admit there could be more practical examples.

```
4. How does yadg handle the non standardized intermediate step (Figure 1)? Would the use of proprietary file formats pose a challenge for FAIR ?
```
`yadg` bypasses it. The data schema as well as the datagram are standardised, with the datagram intended for use with electronic lab book software (archiving) or other python packages, such as the sister tool [dgpost](https://dgbowl.github.io/dgbowl) for postprocessing.

```
5. Kindly provide a brief statement in the article regarding how this software compares to other commonly-used packages? (JOSS requirement)
```
I believe this requirement should already be satisfied by the discussion of the features of `yadg` in the context of units and uncertainties (astropy, pint, uncertainties), or the domain-specific features (PyMS and Aston for chromatography, galvani for electrochemistry). If the reviewer is aware of any omissions, I am happy to investigate and discuss them in the paper.
