+++ paper.bib
@@ -38,14 +38,15 @@
 }
 
 @inproceedings{Draxl2020,
-  title = {{FAIRmat}},
-  booktitle = {{2. NFDI Conference}},
+  title = {{FAIRmat [Presentation]}},
+  booktitle = {{2. NFDI Conference, Online}},
   author = {Draxl, Claudia},
   year = {2020},
   month = jul,
   publisher = {{DFG}},
   address = {{Bonn}},
   langid = {german},
+  howpublished = {https://www.dfg.de/download/pdf/foerderung/programme/nfdi/nfdi_konferenz_2020/fairmat_abstract.pdf},
   file = {C\:\\Users\\krpe\\Zotero\\storage\\5AXGYX4G\\fairmat_abstract.pdf}
 }
 
