+++ paper.md
@@ -41,7 +41,10 @@ An example laboratory workflow is shown in \autoref{fig:workflow}. The devices i
 However, generally it is possible to semantically define a set of raw data related to an experiment, for example by placing them in a folder. Then, a hierarchical description (`schema`) of the data in that folder can be created, containing information about the file types, quantities, their locations, etc. The `yadg` package can be used for reproducbile processing of such `schema` files into a standardised, timestamped, versioned, annotated, and validated `datagram`, which can be directly uploaded to the ELN or used in postprocessing. The file parsers in `yadg` are written in a modular fashion, grouped by scientific application (e.g. chromatography, electrochemistry, etc.), and are fully transparent to the user regardless of the raw file format or device vendor. Additionally, the same `schema` can be used to process different experimental folders using the `preset` functionality of `yadg`.
 
 # Features
-Here, we present the revised `yadg-4.0`. Compared to the previous version of `yadg` [@Kraus2021x], the three main user-facing changes in `yadg-4.0` are the inclusion of units and uncertainties in the raw data, the enhanced chromatogram parser functionality, and the new electrochemistry parser. For a more detailed list of changes, please see the [release notes](https://dgbowl.github.io/yadg/version.html#yadg-version-4-0-0). 
+Here, we present the revised `yadg-4.0`. Compared to the previous version of `yadg` [@Kraus2021x], the three main user-facing changes in `yadg-4.0` are the inclusion of units and uncertainties in the raw data, the enhanced chromatogram parser functionality, and the new electrochemistry parser. For a more detailed list of changes, please see the [release notes](https://dgbowl.github.io/yadg/4.0.0/version.html#yadg-version-4-0-0). 
+
+## Usage
+An interactive, Binder-ready Jupyter notebook, illustrating the installation and usage of `yadg-4.0`, is included on Zenodo under [DOI: 10.5281/zenodo.6351210](https://doi.org/10.5281/zenodo.6351210). The Binder-ready notebook can be launched by following [this link](https://mybinder.org/v2/zenodo/10.5281/zenodo.6351210/?labpath=index.ipynb).
 
 ## Units and uncertainties
 In contrast with the previous version of the tool, raw data is now retained in the `datagram` and is annotated by units and measurement uncertainties, making `datagrams` suitable for archiving in an ELN. While units are often present in the raw data files, the uncertainties often have to be determined from instrument specifications. This is a particularly important feature, as most data scientific packages in the Python ecosystem either support annotating array or tabular data by units ([`astropy`](https://www.astropy.org/) [@astropy;@TheAstropyCollaboration2018] or [`pint`](https://pint.readthedocs.io/en/stable/) [@pint]), or uncertainties ([`unumpy`](https://pythonhosted.org/uncertainties/numpy_guide.html) from the [`uncertainties`](https://pythonhosted.org/uncertainties/index.html) package  [@uncertainties]); the combination of both units and uncertainties is quite rare, especially with non-tabular data.
