### 2022-03-14 f011873
Revert "Fix reference, add test paper build."
This reverts commit 8e126a9f5db074fd73ebe763ca869b199d0696f5.
