+++ paper.bib
@@ -43,7 +43,7 @@
   author = {Draxl, Claudia},
   year = {2020},
   month = jul,
-  publisher = {{Deutsche Forschungsgemeinschaft}},
+  publisher = {{DFG}},
   address = {{Bonn}},
   langid = {german},
   howpublished = {https://www.dfg.de/download/pdf/foerderung/programme/nfdi/nfdi_konferenz_2020/fairmat_abstract.pdf},
