+++ paper.bib
@@ -37,13 +37,13 @@
   file = {C\:\\Users\\krpe\\Zotero\\storage\\UMSVX4B7\\GeCatS_Whitepaper 2019_engl_ezl.pdf}
 }
 
-@inproceedings{Draxl2020,
+@book{Draxl2020,
   title = {{FAIRmat [Presentation]}},
   booktitle = {{2. NFDI Conference, Online}},
   author = {Draxl, Claudia},
   year = {2020},
   month = jul,
-  publisher = {{DFG}},
+  publisher = {{Deutsche Forschungsgemeinschaft}},
   address = {{Bonn}},
   langid = {german},
   howpublished = {https://www.dfg.de/download/pdf/foerderung/programme/nfdi/nfdi_konferenz_2020/fairmat_abstract.pdf},
