+++ paper.bib
@@ -37,17 +37,22 @@
   file = {C\:\\Users\\krpe\\Zotero\\storage\\UMSVX4B7\\GeCatS_Whitepaper 2019_engl_ezl.pdf}
 }
 
-@book{Draxl2020,
+@inproceedings{Draxl2020,
   title = {{FAIRmat [Presentation]}},
-  booktitle = {{2. NFDI Conference, Online}},
   author = {Draxl, Claudia},
-  year = {2020},
-  month = jul,
-  publisher = {{Deutsche Forschungsgemeinschaft}},
-  address = {{Bonn}},
+  crossref = {nfdiconf:20},
   langid = {german},
   howpublished = {https://www.dfg.de/download/pdf/foerderung/programme/nfdi/nfdi_konferenz_2020/fairmat_abstract.pdf},
   file = {C\:\\Users\\krpe\\Zotero\\storage\\5AXGYX4G\\fairmat_abstract.pdf}
+}
+
+@proceedings{nfdiconf:20,
+  title = {{2. NFDI Conference}},
+  publisher = {{Deutsche Forschungsgemeinschaft}},
+  venue = {online},
+  month = jul,
+  year = {2020},
+  langid = german
 }
 
 @techreport{FAIRmat,
