### 2022-03-14 #4166 @PeterKraus
Alright, I'm losing my mind trying to get the conference proceedings formatted properly for https://github.com/dgbowl/yadg/issues/45. @galessiorob, @kyleniemeyer does either of you know how to do this? The following bibtex doesn't seem to work well:

```
@inproceedings{Draxl2020,
  title = {{FAIRmat [Presentation]}},
  author = {Draxl, Claudia},
  crossref = {nfdiconf:20},
  langid = {german},
  howpublished = {https://www.dfg.de/download/pdf/foerderung/programme/nfdi/nfdi_konferenz_2020/fairmat_abstract.pdf},
  file = {C\:\\Users\\krpe\\Zotero\\storage\\5AXGYX4G\\fairmat_abstract.pdf}
}

@proceedings{nfdiconf:20,
  title = {{2. NFDI Conference}},
  publisher = {{Deutsche Forschungsgemeinschaft}},
  venue = {online},
  month = jul,
  year = {2020},
  langid = german
}
```

The title of the conference gets pulled out properly, but `publisher` and `howpublished` do not.
