+++ paper.bib
@@ -37,22 +37,15 @@
   file = {C\:\\Users\\krpe\\Zotero\\storage\\UMSVX4B7\\GeCatS_Whitepaper 2019_engl_ezl.pdf}
 }
 
-@inproceedings{Draxl2020,
-  title = {{FAIRmat [Presentation]}},
-  author = {Draxl, Claudia},
-  crossref = {nfdiconf:20},
-  langid = {german},
-  howpublished = {https://www.dfg.de/download/pdf/foerderung/programme/nfdi/nfdi_konferenz_2020/fairmat_abstract.pdf},
-  file = {C\:\\Users\\krpe\\Zotero\\storage\\5AXGYX4G\\fairmat_abstract.pdf}
-}
-
-@proceedings{nfdiconf:20,
-  title = {{2. NFDI Conference}},
+@unpublished{Draxl2020,
+  type = {Conference},
   publisher = {{Deutsche Forschungsgemeinschaft}},
-  venue = {online},
-  month = jul,
+  title = {FAIRmat},
+  author = {Draxl, C.},
+  institution = {{2. NFDI Conference, online, Deutsche Forschungsgemeinschaft}},
   year = {2020},
-  langid = german
+  month = jul,
+  day = 8
 }
 
 @techreport{FAIRmat,
