+++ paper.bib
@@ -38,11 +38,11 @@
 }
 
 @unpublished{Draxl2020,
-  type = {Conference},
+  type = {Conference paper},
   publisher = {{Deutsche Forschungsgemeinschaft}},
   title = {FAIRmat},
   author = {Draxl, C.},
-  institution = {{2. NFDI Conference, online, Deutsche Forschungsgemeinschaft}},
+  institution = {{2. NFDI Conference, online}},
   year = {2020},
   month = jul,
   day = 8
