+++ paper.bib
@@ -39,7 +39,7 @@
 
 @unpublished{Draxl2020,
   type = {Conference paper},
-  publisher = {{Deutsche Forschungsgemeinschaft}},
+  publisher = {{Deutsche Forschungsgemeinschaft (DFG)}},
   title = {FAIRmat},
   author = {Draxl, C.},
   institution = {{2. NFDI Conference, online}},
