### 2022-03-14 f0c1b6f
JOSS review issues. (#44)
* install instructions in readme.

* Changes to paper.

* Fix reference, add test paper build.

* Build the paper properly.

* Revert "Fix reference, add test paper build."

This reverts commit 8e126a9f5db074fd73ebe763ca869b199d0696f5.

* round 3

* Re-add the paper workflow file.

* Let's try this one

* attempt 4

* Final fix to Draxl.

* remove paper rebuild from PR
