+++ paper.bib
@@ -37,16 +37,15 @@
   file = {C\:\\Users\\krpe\\Zotero\\storage\\UMSVX4B7\\GeCatS_Whitepaper 2019_engl_ezl.pdf}
 }
 
-@inproceedings{Draxl2020,
-  title = {{FAIRmat}},
-  booktitle = {{2. NFDI Conference}},
-  author = {Draxl, Claudia},
+@unpublished{Draxl2020,
+  type = {Conference paper},
+  publisher = {{Deutsche Forschungsgemeinschaft (DFG)}},
+  title = {FAIRmat},
+  author = {Draxl, C.},
+  institution = {{2. NFDI Conference, online}},
   year = {2020},
   month = jul,
-  publisher = {{DFG}},
-  address = {{Bonn}},
-  langid = {german},
-  file = {C\:\\Users\\krpe\\Zotero\\storage\\5AXGYX4G\\fairmat_abstract.pdf}
+  day = 8
 }
 
 @techreport{FAIRmat,
