### 2022-03-14 #4166 @PeterKraus
@pythonpanda2 I think there are a couple unticked/unchecked items wrt. documentation. The testing/CI is done via github actions, and the example usage is now included in a short section in the paper, and it will be added to the docs. See:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6351210.svg)](https://doi.org/10.5281/zenodo.6351210)
