### 2022-03-15 987fe27
Minor fixes (#47)
* Docs changes, gitignore, remove gitlab.

* Change version to 4.1.0rc3
