+++ .gitlab-ci.yml
@@ -1,81 +1,0 @@
-# This file is a template, and might need editing before it works on your project.
-# To contribute improvements to CI/CD templates, please follow the Development guide at:
-# https://docs.gitlab.com/ee/development/cicd/templates.html
-# This specific template is located at:
-# https://gitlab.com/gitlab-org/gitlab/-/blob/master/lib/gitlab/ci/templates/Python.gitlab-ci.yml
-
-# Official language image. Look for the different tagged releases at:
-# https://hub.docker.com/r/library/python/tags/
-image: python:3.9-slim-bullseye
-
-stages:
-  - build
-  - test
-  - pages
-
-# Change pip's cache directory to be inside the project directory since we can
-# only cache local items.
-#variables:
-#  PIP_CACHE_DIR: "$CI_PROJECT_DIR/.cache/pip"
-
-# Pip's cache doesn't store the python packages
-# https://pip.pypa.io/en/stable/reference/pip_install/#caching
-#
-# If you want to also cache the installed packages, you have to install
-# them in a virtualenv and cache it as well.
-#cache:
-#  paths:
-#    - .cache/pip
-#    - venv/
-
-before_script:
-  - python -V  # Print out python version for debugging
-  - pip install virtualenv 
-  - virtualenv venv
-  - source venv/bin/activate
-
-build:
-  stage: build
-  only:
-    - master
-    - merge_requests
-  script:
-    - python setup.py bdist_wheel
-  artifacts:
-    paths:
-      - dist/*.whl
-
-test:
-  stage: test
-  only:
-    - master
-    - merge_requests
-  needs: [build]
-  script:
-    - pip install dist/*
-    - pytest
-
-pages-master:
-  stage: pages
-  only:
-    - master
-  needs: [build]
-  script:
-    - pip install dist/*
-    - pip install sphinx sphinx-rtd-theme sphinx-autodoc-typehints
-    - sphinx-apidoc -o docs/source src/yadg -T -M
-    - sphinx-build -W -b html docs/source public
-  artifacts:
-    paths:
-      - public
-
-pages-mr:
-  stage: pages
-  only:
-    - merge_requests
-  needs: [build]
-  script:
-    - pip install dist/*
-    - pip install sphinx sphinx-rtd-theme sphinx-autodoc-typehints
-    - sphinx-apidoc -o docs/source src/yadg -T -M
-    - sphinx-build -W -b html docs/source public
