+++ tests/test_dummy.py
@@ -2,6 +2,8 @@ import pytest
 import os
 import json
 import yadg.core
+from dgbowl_schemas import Dataschema
+from pydantic import ValidationError
 
 from tests.schemas import dummy_1, dummy_2, dummy_3, dummy_4, dummy_5
 from tests.schemas import fail_1, fail_2, fail_3, fail_4, fail_5
@@ -21,8 +23,8 @@ from tests.schemas import exclude_1
 )
 def test_datagram_from_schema_dict(inp_dict, l_dg, l_res, datadir):
     os.chdir(datadir)
-    assert yadg.core.validators.validate_schema(inp_dict), "invalid schema format"
-    ret = yadg.core.process_schema(inp_dict)
+    ds = Dataschema(**inp_dict)
+    ret = yadg.core.process_schema(ds)
     assert yadg.core.validators.validate_datagram(ret), "invalid datagram format"
     assert len(ret["steps"]) == l_dg, "wrong number of steps"
     if l_dg > 0:
@@ -45,8 +47,8 @@ def test_datagram_from_schema_file(inp_fn, ts, datadir):
     jsonpath = datadir.join(inp_fn)
     with open(jsonpath, "r") as infile:
         schema = json.load(infile)
-    assert yadg.core.validators.validate_schema(schema), "invalid schema format"
-    ret = yadg.core.process_schema(schema)
+    ds = Dataschema(**schema)
+    ret = yadg.core.process_schema(ds)
     assert yadg.core.validators.validate_datagram(ret), "invalid datagram format"
     assert len(ret["steps"]) == ts["nsteps"], "wrong number of steps"
     assert (
@@ -58,14 +60,14 @@ def test_datagram_from_schema_file(inp_fn, ts, datadir):
 @pytest.mark.parametrize(
     "inp_dict, expr",
     [
-        (fail_1, r"'parser'"),
-        (fail_2, r"'dumm'"),
-        (fail_3, r"2 were provided"),
-        (fail_4, r"0 were provided"),
-        (fail_5, r"'key'"),
+        (fail_1, r"parser"),
+        (fail_2, r"name 'dumm' provided as a 'parser'"),
+        (fail_3, r"Specifying multiple arguments"),
+        (fail_4, r"Either 'files' or 'folders'"),
+        (fail_5, r"extra fields not permitted"),
     ],
 )
 def test_schema_validator(inp_dict, expr, datadir):
     os.chdir(datadir)
-    with pytest.raises(AssertionError, match=expr):
-        assert yadg.core.validators.validate_schema(inp_dict)
+    with pytest.raises(ValidationError, match=expr):
+        assert Dataschema(**inp_dict)
