+++ tests/test_preset.py
@@ -2,6 +2,7 @@ import pytest
 import os
 import json
 from tests.utils import standard_datagram_test
+from dgbowl_schemas import Dataschema
 
 import yadg.dgutils
 import yadg.core
@@ -21,7 +22,7 @@ def test_preset(input, ts, datadir):
         preset = json.load(infile)
 
     schema = yadg.dgutils.schema_from_preset(preset, input)
-    assert yadg.core.validate_schema(schema)
+    ds = Dataschema(**schema)
 
-    ret = yadg.core.process_schema(schema)
+    ret = yadg.core.process_schema(ds)
     standard_datagram_test(ret, ts)
