### 2022-03-16 #4166 @galessiorob
Hi @PeterKraus,

We're almost at the finish line! I put together a few edit suggestions and typo corrections in this [PR](https://github.com/dgbowl/yadg/pull/49). Please merge as much as you like, and after that, we'll review once more and start the publishing process.
