### 2022-03-16 ee8c5c8
Add some suggestions to the paper (#49)
* Add some suggestions to the paper

* Period behind quotes.

Co-authored-by: krpe <peter.kraus@empa.ch>
