+++ paper.md
@@ -26,22 +26,22 @@ bibliography: paper.bib
 
 # Summary
 
-The management of scientific data is a key aspect of modern data science. Four simple guiding principles that are combined under the FAIR data moniker define the current "gold standard" in data quality: the data has to be **F**indable by anyone, **A**ccessible without barriers,  **I**nteroperable with other programs, and **R**eusable after analysis [@Wilkinson2016]. Yet, many scientific data formats do not conform to these principles. This is especially true for proprietary formats, often associated with expensive lab instrumentation. The `yadg` package helps to resolve this issue by parsing raw data files into a standardised, annotated, and timestamped format readable by both humans and machines. Various raw data formats are supported, including chromatograms, electrochemical cycling protocols, reflection coefficient traces, spectroscopic data, and tabulated data. The parsed files include information about data provenance, units of measure, and experimental uncertainties by default. Finally, several common data processing steps, such as applying calibration functions, integration of chromatographic traces, or fitting of reflection coefficients, are available in `yadg`. 
+The management of scientific data is a crucial aspect of modern data science. Four simple guiding principles combined under the FAIR data moniker define the current "gold standard" in data quality: the data has to be **F**indable by anyone, **A**ccessible without barriers,  **I**nteroperable with other programs, and **R**eusable after analysis [@Wilkinson2016]. Yet, many scientific data formats do not conform to these principles. This is especially true for proprietary formats, often associated with expensive lab instrumentation. The `yadg` package helps to resolve this issue by parsing raw data files into a standardised, annotated and timestamped format readable by both humans and machines. Various raw data formats are supported, including chromatograms, electrochemical cycling protocols, reflection coefficient traces, spectroscopic data, and tabulated data. The parsed files include information about data provenance, units of measure, and experimental uncertainties by default. Finally, several common data processing steps, such as applying calibration functions, integration of chromatographic traces, or fitting of reflection coefficients, are available in `yadg`. 
 
 # Statement of need
 
-From the point of view of catalytic chemistry, digitalisation is currently a "hot topic". For example, the leading German and Swiss academic bodies in the field of catalysis consider digitalisation a key task for the current decade. The German Catalysis Society has published a detailed roadmap [@Demtroder2019], and large consortia have been formed (eg. FAIRmat [@Draxl2020] in Germany or NCCR Catalysis in Switzerland [@NCCRcat]) to advance this issue. 
+From the point of view of catalytic chemistry, digitalisation is currently a "hot topic". For example, the leading German and Swiss academic bodies in the field of catalysis consider digitalisation an essential task for the current decade. The German Catalysis Society has published a detailed roadmap [@Demtroder2019], and large consortia have been formed (eg. FAIRmat [@Draxl2020] in Germany or NCCR Catalysis in Switzerland [@NCCRcat]) to advance this issue. 
 
-However, this change will not happen overnight. One approach for kick-starting this transition is to make the currently existing "small data" available according to the FAIR principles [@Mendes2021], effectively defining a standard for "big data" by superposition of current best practices. Another approach, more specific to catalysis, is that of standardisation of testing and characterisation protocols [@Trunschke2020], which will necessarily lead to a standardised data content. The `yadg` package presented here has been conceived to combine both of those approaches, and when combined with a lab automation software and data post-processing tools, forms a robust data-scientific pipeline [@Kraus2021x].
[…]
