### 2022-04-05 #4166 @galessiorob
@PeterKraus 🙈 I am very sorry I missed this question. I apologize for the delay and hope you have more upcoming talks where you can point the audience to your paper. I'm a Chemical Engineer and this paper is awesome! I'm sure it'll be very well received.

The next thing we need from you is to deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing.

@aozorahime @pythonpanda2 @1mikegrn Thank you so much for lending your expertise and reviewing this paper 🚀
