### 2022-04-05 #4166 @PeterKraus
Hi @galessiorob, no worries about the delay. Glad you like the paper/software! I have uploaded all published versions of yadg to https://doi.org/10.5281/zenodo.5894822 (with version 4.0.0 being https://doi.org/10.5281/zenodo.6413504). We're working on getting v4.1.0 out, which will have some of the suggestions from reviewers merged in, but it's not ready yet as we need to finish a couple of features. Do I need to add the DOIs to the paper, or are the links here enough?
