### 2022-04-06 c764f18
Fix references in bibtex file (#51)
* Bibtex fix #1

* enable test build paper
