+++ paper.bib
@@ -73,21 +73,19 @@
   volume = {585},
   number = {7825},
   pages = {357--362},
-  issn = {0028-0836, 1476-4687},
+  issn = {1476-4687},
   doi = {10.1038/s41586-020-2649-2},
   abstract = {Abstract                            Array programming provides a powerful, compact and expressive syntax for accessing, manipulating and operating on data in vectors, matrices and higher-dimensional arrays. NumPy is the primary array programming library for the Python language. It has an essential role in research analysis pipelines in fields as diverse as physics, chemistry, astronomy, geoscience, biology, psychology, materials science, engineering, finance and economics. For example, in astronomy, NumPy was an important part of the software stack used in the discovery of gravitational waves               1               and in the first imaging of a black hole               2               . Here we review how a few fundamental array concepts lead to a simple and powerful programming paradigm for organizing, exploring and analysing scientific data. NumPy is the foundation upon which the scientific Python ecosystem is constructed. It is so pervasive that several projects, targeting audiences with specialized needs, have developed their own NumPy-like interfaces and array objects. Owing to its central position in the ecosystem, NumPy increasingly acts as an interoperability layer between such array computation libraries and, together with its application programming interface (API), provides a flexible framework to support the next decade of scientific and industrial analysis.},
   langid = {english},
   file = {C\:\\Users\\krpe\\Zotero\\storage\\8Y2QXEPE\\Harris et al. - 2020 - Array programming with NumPy.pdf}
 }
 
-@unpublished{Kraus2021x,
-  type = {Preprint},
+@article{Kraus2021x,
   title = {Towards Automation of Operando Experiments: {{A}} Case Study in Contactless Conductivity Measurements},
-  shorttitle = {Towards Automation of Operando Experiments},
   author = {Kraus, Peter and Wolf, Elisabeth and Prinz, Charlotte and Bellini, Giulia and Trunschke, Annette and Schl{\"o}gl, Robert},
-  year = {2021},
-  month = oct,
-  doi = {10.33774/chemrxiv-2021-mh17g},
+  year = {2022},
+  doi = {10.1039/D1DD00029B},
+  journal = {Digital Discovery},
   abstract = {Automation of experiments is a key component on the path of digitalisation in catalysis and related sciences. Here we present the lessons learned and caveats avoided during the automation of our contactless conductivity measurement set-up, capable of operando measurement of catalytic samples. We briefly discuss the motivation behind the work, the technical groundwork required, and the philosophy guiding our design. The main body of this work is dedicated to the detailing of the implementation of the automation, data structures, as well as the modular data processing pipeline. The open-source toolset developed as part of this work allows us to carry out unattended and reproducible experiments, as well as post-process data according to current best practice. This process is illustrated by implementing two routine sample protocols, one of which was included in the Handbook of Catalysis, providing several case studies showing the benefits of such automation, including increased throughput and higher data quality. The datasets included as part of this work contain catalytic and operando conductivity data, and are self-consistent, annotated with metadata, and are available on a public repository in a machine-readable form. We hope the datasets as well as the tools and workflows developed as part of this work will be an useful guide on the path towards automation and digital catalysis.},
   copyright = {All rights reserved},
   langid = {english},
@@ -104,7 +102,7 @@
   volume = {13},
   number = {3},
   pages = {836--850},
-  issn = {1867-3880, 1867-3899},
+  issn = {1867-3880},
   doi = {10.1002/cctc.202001132},
   langid = {english},
   file = {C\:\\Users\\krpe\\Zotero\\storage\\EJLB2DWP\\Mendes et al. - 2021 - Open Data in Catalysis From Today's Big Picture t.pdf}
@@ -190,7 +188,7 @@
   journal = {Topics in Catalysis},
[…]
