### 2022-04-08 #4166 @kyleniemeyer
@galessiorob sorry, I'm not sure either! The paper header looks alright to me. @openjournals/dev any ideas what's going on here? The paper author/affiliations block is:
```yaml
authors:
  - name: Peter Kraus^[corresponding author]
    orcid: 0000-0002-4359-5003
    affiliation: 1
  - name: Nicolas Vetsch
    affiliation: 1
  - name: Corsin Battaglia
    affiliation: 1
affiliations:
 - name: |
      Empa - Swiss Federal Laboratories for Materials Science and Technology,
      Überlandstrasse 129, 8600 Dübendorf
      Switzerland
   index: 1
```
