### 2022-04-08 #4166 @danielskatz
maybe change the affil entry to a single line:
```
 - name: Empa - Swiss Federal Laboratories for Materials Science and Technology, Überlandstrasse 129, 8600 Dübendorf, Switzerland
```
?
I'm not sure about the pipe character or the multiline affiliation
