+++ paper.md
@@ -35,7 +35,7 @@ However, this change will not happen overnight. One approach for kick-starting t
 
 An example laboratory workflow is shown in \autoref{fig:workflow}. The devices in the laboratory (eg. chromatographs, flow meters, potentiostats, temperature controllers, or spectrometers) create raw data files that are different in shape, size, and number. If one wishes to store this raw data in an electronic lab notebook (ELN) software in a FAIR way or use it in post-processing, the raw data must be parsed. The intermediate parsing step is often not standardised: in the case of chromatography, the data parsing may be specific to the particular data files as opposed to file types, and in the case of flow and temperature data the values are often entered manually and hence prone to human error. 
 
-However, it is generally possible to semantically define a set of raw data related to an experiment by placing them in a folder. Then, a hierarchical description (`schema`) of the data in that folder can be created, containing information about the file types, quantities, their locations, etc. The `yadg` package can be used for reproducible processing of such `schema` files into a standardised, timestamped, versioned, annotated, and validated `datagram`, which can be directly uploaded to the ELN or used in post-processing. The file parsers in `yadg` are written in a modular fashion, grouped by scientific application (e.g. chromatography, electrochemistry, etc.), and are fully transparent to the user regardless of the raw file format or device vendor. Additionally, the same `schema` can be used to process different experimental folders using the `preset` functionality of `yadg`.
+However, it is generally possible to semantically define a set of raw data related to an experiment by placing them in a folder. Then, a hierarchical description (`schema`) of the data in that folder can be created, containing information about the file types, quantities, their locations, etc. The `yadg` package can be used for reproducible processing of such `schema` files into a standardised, timestamped, versioned, annotated, and validated `datagram`, which can be directly uploaded to the ELN or used in post-processing. The file parsers in `yadg` are written in a modular fashion, grouped by scientific application (e.g., chromatography, electrochemistry), and are fully transparent to the user regardless of the raw file format or device vendor. Additionally, the same `schema` can be used to process different experimental folders using the `preset` functionality of `yadg`.
 
 # Features
 Here, we present the revised `yadg-4.0`. Compared to the previous version of `yadg` [@Kraus2021x], the three main user-facing changes in `yadg-4.0` include units and uncertainties in the raw data, the enhanced chromatogram parser functionality, and the new electrochemistry parser. For a more detailed list of changes, please see the [release notes](https://dgbowl.github.io/yadg/4.0.0/version.html#yadg-version-4-0-0). 
@@ -47,7 +47,7 @@ An interactive, Binder-ready Jupyter notebook, illustrating the installation and
 In contrast with the previous version of the tool, raw data is now retained in the `datagram` and is annotated by units and measurement uncertainties, making `datagrams` suitable for archiving in an ELN. While units are often present in the raw data files, the uncertainties often have to be determined from instrument specifications. This is a particularly important feature, as most data scientific packages in the Python ecosystem either support annotating array or tabular data by units ([`astropy`](https://www.astropy.org/) [@astropy;@TheAstropyCollaboration2018] or [`pint`](https://pint.readthedocs.io/en/stable/) [@pint]), or uncertainties ([`unumpy`](https://pythonhosted.org/uncertainties/numpy_guide.html) from the [`uncertainties`](https://pythonhosted.org/uncertainties/index.html) package  [@uncertainties]); the combination of both units and uncertainties is quite rare, especially with non-tabular data.
 
 ## Chromatography
[…]
